renamed autologin plugin to auth, added comment step
