+++ lib/step/base.js
@@ -2,6 +2,7 @@ const color = require('chalk')
 const Secret = require('../secret')
 const { getCurrentTimeout } = require('../timeout')
 const { ucfirst, humanizeString, serializeError } = require('../utils')
+const recordStep = require('./record')
 
 const STACK_LINE = 5
 
@@ -49,6 +50,10 @@ class Step {
 
   run() {
     throw new Error('Not implemented')
+  }
+
+  addToRecorder(args) {
+    return recordStep(this, args)
   }
 
   /**
