+++ lib/step/comment.js
@@ -1,0 +1,10 @@
+const FuncStep = require('./func')
+
+class CommentStep extends FuncStep {
+  constructor(name, comment) {
+    super(name)
+    this.fn = () => {}
+  }
+}
+
+module.exports = CommentStep
